+++ Source/FreeTrainSimulator.Models/Settings/ProfileDispatcherSettingsModel.cs
@@ -7,7 +7,7 @@ namespace FreeTrainSimulator.Models.Settings
 {
     [MemoryPackable(GenerateType.VersionTolerant, SerializeLayout.Sequential)]
     [ModelResolver("", ".dispatchersettings")]
-    public partial record ProfileDispatcherSettingsModel : ProfileSettingsModelBase
+    public sealed partial record ProfileDispatcherSettingsModel : ProfileSettingsModelBase
     {
         public override ProfileModel Parent => base.Parent as ProfileModel;