+++ WHATSNEW.md
@@ -2,9 +2,9 @@
 
 ## Features
 
-- Make keyboard-input settings profile dependent, allowing to quickly change between different keyboard profiles
-- Profile-dependent Dispatcher
-- Initialize Testing from user profile settings
+- Profile-dependent keyboard input settings, allowing to quickly change between different keyboard profiles
+- Profile-dependent Dispatcher settings
+- Initialize Testing with user profile settings
 - Loading Language settings from user profile
 - Log files are stored in the user's application data folder (i.e. C:\Users\USERNAME\AppData\Roaming\Free Train Simulator\Logs). 
 - New button on the Menu-toolbar opens the Log file folder for quick access
@@ -19,3 +19,7 @@
 - Fixing a race condition in FolderStructure
 
 ## Maintenance
+
+## Known Issues
+
+- Toolbox settings are also stored with the current profile, but not copied when cloning profiles